requirements.txt
flask
pyTelegramBotAPI==4.7.0
python-dotenv
